feat(display): Aero and Liquid glass monitor faces
Two new monitor faces built on a shared glass compositor: STYLE_GLASS_TILES
(the Tiles grid on polished Vista-style panes) and STYLE_GLASS_DUO (the Duo
bands and pills as edge-lit capsules).

The panel has no framebuffer to blur, and does not need one: a smooth vertical
gradient, blurred, is still itself. The backdrop is a generated vertical ramp,
which makes every pane colour a pure function of the panel row - so a pane can
be composed in sprite-sized windows that join seamlessly and never has to read
the panel back.

Compositor:
- Blending runs at 8 bits per channel and is dithered on the way down to
  RGB565 with a centred 8x8 Bayer threshold. Blending in 565 space bands
  badly; a 4x4 cell reads as a visible checkerboard at this pixel pitch.
- Rounded corners, rim and outer edge are coverage-blended from a sub-pixel
  inset, so nothing on these faces has a stepped edge.
- Charts get monotone-clamped Catmull-Rom interpolation, an antialiased
  stroke and a depth-faded area fill that blends into the pane instead of
  punching a flat block through it.
- Charts glide: the plot scrolls by a fraction of a sample between packets
  and the newest reading slides in over the interval, so the series moves
  continuously instead of stepping once a second. The reading itself is not
  eased - a stats readout must not show numbers the sensor never reported.

Warn moves to the rim. Retinting the pane body left the chart inside looking
like a foreign rectangle, since the plot keeps its identity hue; every other
face here keeps its card identity-coloured too.

Byte order: LovyanGFX stores 16-bit sprite pixels in the panel's order, so
raw buffer writes bypass its conversion. The order is probed once rather than
assumed, and glassBackdrop states its pushImage expectation explicitly.

Frame budget measured, not guessed: /api/status now reports frame_us and
frame_max_us. Glass Tiles at eight metrics costs ~42 ms on the S3, so the
animated cadence is 10 Hz - the chart only scrolls ~1.7 px/second, and a
faster rate would burn CPU redrawing an unchanged picture.

TouchSettings v2 widens styleMask to 16 bits for styles 7 and 8, with a v1
read path so a configured pad survives the update.

All five envs build; both faces verified on the esp32s3 at 240x240.

+++ src/touch_button.cpp
@@ -65,7 +65,7 @@ static void recordAction(const char* action) {
 }
 
 static void cycleStyle(int8_t direction) {
-  uint8_t mask = touchSettings.styleMask & STYLE_ACTIVE_MASK;
+  uint16_t mask = touchSettings.styleMask & STYLE_ACTIVE_MASK;
   if (mask == 0) mask = 1u << STYLE_DEFAULT;
 
   for (uint8_t step = 1; step <= STYLE_COUNT; step++) {
